refactor: remove unused flag type and restructure query state lib
